Inline codegen_mono_item
